Refactor CustomObjects and Resources
There are now CustomObjects, and all BO2 are a CustomObject. UseWorld
and UseBiome are also CustomObjects. Even all trees are now a
CustomObject.

All resources now inherit Resource. All settings don't have to be stored
in a generic Resource object, but they are stored in the resource
itself. This allow for more flexibility.

Plugin developers can now add their own CustomObjects and Resources.