Properly fix ice and snow; refactoring
The previous fix moved the snow placement 8 blocks. The correct way to
fix it would actually be to leave the snow alone, but to move the trees
and CustomObjects.

As @oloflarsson suggested, all fields in BukkitWorld are now camelCase
instead of CamelCase. It is a huge task to convert every field and
method in TC to camelCase, but it will slowly happen.