Change block.data syntax to block:data
Was harder than it sounds, because ConfigFile kept thinking that
something with a : in it is a setting instead of a resource. Now it is
smarter and knows the difference between ReplaceBlocks and Ore.

Other changes:
- Moved most methods from ConfigFunction to the new class StringHelper,
so that the rest of the config file can also benefit from them
(currently unused).
- Deleted TxT.java, and moved it's most useful methods to the
StringHelper.java.