Fix UnderWaterOre, add debug code, ...
... add tripwire rotation, rename replaceBiomesLate to replaceBiomes.

There is something wrong with setBlock and the meta data. I found out
that some BOB blocks are always rotating incorrectly. I thought that the
bug was in BlockHelper, but this code proves that it is outside
TerrainControl. I wonder if TerrainControl is doing something wrong.
Anyways, I need to find a workaround.