"Fix" /reload
Cleaning up old parts after a /reload of TC is not really possible for
some reason. So we simply don't enable the new parts, because they won't
work with the old parts correctly. This means that custom colors,
commands, custom sapling growing won't work after a /reload. Terrain
generationi still works however, but the configs aren't reloaded (and
cannot be reloaded after a /reload because the commands don't work).

Of couse, a big warning is shown after a /reload to notify server admins
that parts of TC won't function correctly.