Use PlayerRegisterChannelEvent to send packet
The custom colors packet is now sent in the PlayerRegisterChannelEvent
instead of one second after the PlayerJoinEvent.

Should cut down CPU usage a bit, as the Bukkit scheduler is no longer
used. This also sends the packet to the client when the client asks for
it, which is better than one second after the player joined.