Remove writeNewLine() in configs
Now a new line is included in writeValue(..). This change removes a lot
of boilerplate code. It also makes it harder to write too few or too
many blank lines.

Sometimes it is still necessary to print out a blank line, for example
in a long comment. You can still use writeComment("") for this - when
the comment is blank, the # at the beginning of the line won't be
printed. (This behaviour of writeComment("") existed before.)