Temporary fix for logger
This is not an ideal fix, some features of the old logger are missing
(like the logger settings in the PluginConfig!), but at least it doesn't
crash anymore.

The logIfLevel methods were removed because it was easier to remove them
than to fix them. The should probably be readded.

In the future the java.util.logging.Level class should be replaced by
org.apache.logging.log4j.Level, but for now it is just translated.