Add the new tree objects
The new trees are not yet added to the default settings, but people can
now use them in  the Tree resource, the Sapling setting and the /tc
spawn command.

Also renamed the Forest tree to Birch.

Also fixed a crash on TerrainMode:Default when spawning a tree using /tc
spawn.