Give ReplaceToBiomeName an offset of 8 blocks
When terrain is populated, 4 chunks are guaranteed to be loaded: (x,z),
(x+1,z), (x,z+1) and (x+1,z+1). Because of this, all populators use an
offset of 8 blocks towards the top left chunk. The area being populated
looks like this:

+--------+--------+ . = no changes in biome for now
|........|........| # = biome is replaced
|....####|####....|
|....####|####....| The top left chunk is saved as chunk 0
+--------+--------+ in the cache, the top right chunk as 1,
|....####|####....| the bottom left as 2 and the bottom
|....####|####....| right chunk as 3.
|........|........|
+--------+--------+

This has the advantage that it is no problem if something like a tree
tries to place something a few blocks outside the area.

Up until now, ReplaceToBiomeName didn't use this area, it replaced the
biomes in the top left chunk. This introduced all kinds of subtle bugs.
The commit fixes that.