Fix more compile errors in Forge, fix NBT logic
When fixing compile error in the Forge version, I discovered a bug in
the Bukkit version. NBT lists tags with other things than compound tags
will now be copied properly.

Metrics isn't updated for Forge 1.7.2, so it has been removed for now.
It can be brought back when an update is out.