Add gradle files, deprecate Maven files
This is more or less forced upon us by ForgeGradle. ForgeModderTools is
broken (problems with inner classes, problems with recompiling), and I
have no idea how and where it needs to be fixed.

Gradle does give us some nice benefits:

* The Terrain Control version now needs to be updated on only one place,
the main build.grade file.
* The build files are a little less verbose.
* Even if you don't have Gradle installed, you can still compile Terrain
Control: a simple Gradle wrapper (gradlew) is included. It automatically
downloads Gradle. It works on all operating systems. gradlew (and gradle
too) require that Java is added to the PATH system variable or that the
JAVA_HOME variable is set.

See also this comment:
https://github.com/Wickth/TerrainControl/commit/58f753d1eeea938bf1f5ddec5217748118ea3951#commitcomment-5227440