Improve /tc biome command
The command now has a new option, -s, to view the biome that is saved to
the world. The command now also displays the biome id, in addition to
the biome name.

The command no longer lists the biome wetness when using the -f flag,
you can just look it up in the BiomeConfig.

The temperature display in -f has been improved, it will now take the
height into account too.