Fix method flow in BiomeConfigFinder
TC won't stop on error anymore, it will just continue.
When it has detected a file is a directory, it will now properly search
in it.

This bug was masked, because when a biome file isn't found, TC just
creates a BiomeConfig instance with the default file name. The
BiomeConfig would then notice that a file was actually there, and read
the file anyway. Only when a) subfolders where used and b) the biome
extension was changed this bug became clear.

Thanks to @Lantoaster for finding this bug.