Implement virtual biomes on Forge
Also made some changes to how the code is structured.

LocalBiome.getCustomId
======================
In the past, we had some ordering of isle biomes based on a "custom id",
which was dependant on the order in which biomes were listed in the
WorldConfig (in CustomBiomes), instead of on the biome id, like most
other things. As the CustomBiomes setting is now ordered automatically
based on id, it makes no sense anymore to leave that code in. The
LocalBiome interface no longer has the getCustomId method.

TCWorldChunkManager.a
=====================
This method doesn't do anything, doesn't override anything and doesn't
need to override anything, as Minecraft provides a spawnable biome list
itself, with exactly the same biomes (except that our method wasn't
updated in ages). The method has been removed.

Less parameters in BukkitBiome.forCustomBiome
=============================================
All those extra parameters we had previously aren't needed at all, we
can just construct the CustomBiome instance ourselves.

TerrainControlEngine.getWorld
=============================
The method without a world name has been removed, this method is part of
the internal Forge implementation and shouldn't be exposed in the Common
part. The Common part is made with multiworld support in mind, the
getWorld method without the world name parameter doesn't belong there.