Don't let virtual biomes overwrite real biomes
For 7e4d03593217d438eff6d2b2e666002653e20ac5 I had to remove the biome
ids from the early biome loading stages. This caused the biomeConfig
load order to become unpredictable (not the terrain generation order).
This exposed a bug: if a virtual biome was registered before its real
counterpart (the one with the same saved id), the virtual biome would
overwrite the real biome in the BiomeBase array. This caused some biomes
to not generate at all.

Thanks to saturnineNL for reporting this bug.