Fix virtual biome ids ending up in the world save files in Forge. Fixes #269
This bug was masked for a long time.. Because of the cast to byte, it
wasn't obvious that the chunk biome array in Forge contained garbage
data.

Previously: shove every biome id, virtual or not, in the world save
files, cast everything to a byte so that it will fit.
Now: actually look up the correct saved id based on the virtual id.