Fix BiomeConfig.moveSettingFromWorld
It previously used Object.equals to determine whether the WorldConfig
has a setting that needs to be moved over.  Unfortunately, double[]
doesn't implement this method as expected: it checks for reference
equality.

The fix is to simply check for the existance of the setting in the
WorldConfig, instead of checking the value.