Move setter to VanillaBiomeGenerator
TCWorldChunkManager now expects a BiomeGenerator in the constructor.
VanillaBiomeGenerator now has a setter for the WorldChunkManager,
instead of the class trying to find it by itself.

The code is not ideal, but the problem is as follows:

* TCWorldChunkManager depends on BiomeGenerator
* VanillaBiomeGenerator depends on (vanilla) WorldChunkManager

There isn't really a way to solve this, as WorldChunkManager has two
roles (both the vanilla biome generator and the biome generator
interface). We have no control over WorldChunkManager, so we have to
keep it like this until Mojang introduces a proper modding API.