Fix ReplacedBlocks when two chunks are populated at once
In 4035249694073ed2b407bc85846cc20b05eedff2, it was made possible for
two chunks to populate at once. This was done to compensate for other
mods poking around in unloaded chunks, causing those chunks to be
populated during population.

The ReplacedBlocks code didn't account for this possibility. This caused
errors when the cache was null. The code has been rewritten a bit to
support this.