Fix biome cache getting corrupted
As usual, it's very difficult to find the cause of a bug. When you have
found it, it is trivial to fix.

Ini this case weird patterns of Mesa sand were sometimes generated by
the surface populator where there should be Plains grass. A Mesa biome
was nearby. I thought this would be a chunk rotation issue, or a case of
using unzoomed biomes instead of the zoomed ones. I was wrong.

When standing on those patches where the surface generator errored out,
sometimes /tc biome would report the surface block that accidentically
generated, sometimes it would report the biome that should have
generated.

I typed /tc biome, it returned Mesa, matching the weird spot on the
ground, but it should have been plains. I went to another patch of mesa
sand, and it returned Plains. Went back to the original spot, and not it
returned Plains.

This pointed to an error in the biome cache. I added debug lines
everywhere, which became so spammy that is wasn't useful at all. I
limited the debug lines to a single chunk (if
(chunkCoord.coordsMatch(0,3) { ...debug lines.. }). When creating a
cache block, the correct biomes were printed. Yet somehow the surface
generator and /tc biome got the wrong ids sometimes.

I added code to display exactly what was returned every time the cache
was accessed. The logs became extremely spammy again, but I noticed that
it returned different biome ids sometime. Something must be modifying
the cache blocks after they are created...

Returning a copy of the interal cache array fixes the problem.