Add hidden PopulateUsingSavedBiomes setting
http://dev.bukkit.org/bukkit-plugins/terrain-control/forum/70486-pre-defined-world-painter-map-problem/

To implement virtual biomes, we had to query the biome generator to be
able to correctly populate virtual biomes. However, some people actually
want biomes to be read from the map files, for example to populate
biomes created by WorldPainter. A hidden setting has been added for
this.

See this wiki page for more information:
https://github.com/Wickth/TerrainControl/wiki/WorldPainter