Replace some calls with DefaultMaterial.AIR
In a0c90789d12092ed9c84a9bc16173d468f481b7c a method was added that
makes checking for air a little faster. It saves exactly one lookup call
to a HashMap (the one to convert DefaultMaterial into a Minecraft
Block). That tiny difference mattered a lot for performance critical
things like chunk generation, where checking for air happens *a lot*.

For consistency all calls in terrain population were changed too. It
doens't give a measurable performance benefit.

Again just for consistency a call in BukkitWorld was changed to mirror
the call in BukkitChunkBuffer (and ForgeWorld and ForgeChunkBuffer).