Don't use biome group id 0
I haven't yet tried to find out why, but the id 0 is not a valid biome
group id, and causes the group to not generate at all. I suspect it's
used as a null value, instead of the often used -1.

Also removed a parameter from the isRoomForMoreGroups method, parameter
was not used.