Update to Minecraft 1.8
To build this file you need to have both Spigot and Spigot-API in your
local Maven repo. Spigot's BuildTools.jar does this automatically for
you. See platforms\bukkit\build.gradle for more information.

Lots of small changes, many are straightforward. A few notable things:
* BukkitMaterialData has been rewritten for the new IBlockData class.
The cache now also supports block data, as it can now use the IBlockData
class as a key. Previously, it used Block as the key.
* The default data value is no longer always 0. For example, pumpkins
have a default data value of 2. The code has been updated to only write
block data if it's not the default for the block.
* Block lookups use a little hack (accessing ChunkSections directly) to
look up the block. This saves creation (and garbage collection) of a ton
of BlockPosition instances.
* Mojang made a variable protected in WorldProvider, but we still need
to access this field. Some reflection (based on field type, not name) is
now used. The ReflectionHelper class has been added for this. The other
use of reflection; the village generator, has been updated to use this
new class.
* Mojang made the world.isStatic variable final. The little hack to fix
torches popping off in BO2/3s has been removed. This means that in the
BO3 the Block(...) to place the wall now needs to be placed before the
Block(...) to place the torch correctly.