Use Spigot dependency from my personal Maven repo
To (hopefully) be able to legally post a Spigot jar, the code has been
stripped from the jar: only class, method and interface signatures
remain.

We used this trick earlier back in the ForgeModderTools days to be able
to distribute a copyrighted JAR. Back then we used it to host a
minecraft.jar.

Tests depending on Spigot code won't run without said code, so they now
disable themselves when needed.