Update Forge version to 1.8
Ocean monuments are already working.

ForgeMaterialData and
ForgeChunkBuffer have been rewritten for the changes made by
Mojang.

The BlockPos refactor caused many methods to be unnamed, so
lots of
func_78485_a in the code. I have added a comment where a proper
name of the method is not immediately clear.