Fix BiomeMode:BeforeGrouped not looking completely like old layout
The actual fix was on line 163 of BiomeGroup.java. The rarity of the
group itself shouldn't be included in the total group rarity....

As always with debugging: it's easy if you know where to look. The bug
was in a completely unexpected place...

Before this branch is merged to master, I want to remove some log
messages and improve the default settings a bit.

Also fixed a potential bug of biome groups created by LayerFactory not
being processed. No idea if it actually matters, but it may bite us in
the future if it isn't fixed now.