Rewrite /tc check to check whether a world is enabled
Previously, it would check the syntax of the config files and output the
results to the console. If the config files were not already created, it
would create them, which is very confusing: it looks like TC was just
enabled for the world, but still isn't .

The command now simply checks whether TC is enabled for the given world.