Remove /tc replace
This command hasn't worked properly for ages: if it doesn't crash the
server, it just randomly skips chunks. It is very dangerous to run this
command, it will corrupt your world. Editing the world files on a live
server is not a good idea. Other programs exist to change the biome ids
in a map, as explained here:

https://github.com/MCTCP/TerrainControl/wiki/Upgrading-an-old-map-to-Minecraft-1.7