Change how BiomeGroups are read; deprecation cleanup
The "fake settings" HotBiomes and ColdBiomes no longer exist in the
WorldStandardValues. The conversion of the old biome lists now happens
in a similar way of how DisableNotchPonds is converted to a resource: in
renameOldSettings. All old settings handling is now located in that
method, no longer are parts of it hidden in readBiomeGroups. Creation of
the default biome groups is now also similar to the creation of the
default resources, except that the createDefaultBiomeGroups method is
placed in the WorldConfig.

In the future, something needs to be done about how default values work,
so that the code is easier to follow.

All deprecated methods and fields have been removed in preperation for
TC 2.7.