Remove LocalMaterialData cache
In Minecraft 1.8, it is quite cheap to create this instance. For
example, ForgeMaterialData only holds a single IBlockData instance.

However, calculating the hashcode is now quite expensive: it needs to
look up the block id and data. The data is calculated from the block
states. So putting LocalMaterialData instances in a hashmap is not such
a good idea anymore.

Therefore the cache has been removed. On my machine, the total time
spent for the cache was about 5 seconds for the initial chunk generation
of a new world, so this is a nice speedup.

Of course, the garbage collector now has got a more work to do, but
still, the initial world generation is now a few seconds faster.

The next bottleneck is the NoiseGeneratorPerlinOctaves.Noise3D method,
but I'm afraid we can't really speed up that.

World.n() also takes 2.3 seconds in total for the intital chunk
generation, but calling this method is required if we want to let the
blocks show up on the client. However, if the chunk is not yet sent to
other players, we may skip the call to this metohd.

BukkitMaterialData.ofIds() takes 1.8 seconds, I wonder if something can
be done to speed that up.