Fix HugeMushroom spawning only one kind of mushroom
This bug only happens on the Spigot version. Basically, at server
startup a kind of mushroom (red or brown) would be chosen and that
mushroom would be spawned every time a HugeMushroom was spawned.

This is caused by a change in the Minecraft WorldGenHugeMushroom class,
which now chooses a random mushroom kind at server startup instead of at
each spawn. This change happened somewhere in between the Minecraft
1.8.0 release and 1.8.8, so that explains why the Forge version didn't
have this bug (the Forge version still uses Minecraft 1.8.0).

This commit now makes HugeMushroom random again, and also adds
HugeRedMushroom and HugeBrownMushroom for people who actually want only
one kind of mushroom.