Winter Wonderland!
This commit adds several new features:

UseTemperatureForSnowHieght: By Default, all snow is 1 layer high.
When this setting is set to true, snow height is determined by biome
temperature and therefore height. For now: A block temp > -.5 yields
a single snow layer. A block temp < -.75 yields max snow layers. All
values in the range -.75 < temp < -.5 are evenly distributed. Defaults
to False. This was in the biomeConfig a couple commits back, it's a
little more user friendly when set on a per-world basis.

BetterSnowFall: By Default, snow falls on the highest block only.
Setting this to true will cause snow to fall through leaves but leave
a little snow on the way. Defaults to False

FullyFreezeLakes: By Default in cold biomes, lakes freeze but only
water exposed to sky is frozen. Setting this to true causes any lake
in a cold biome with at least one block exposed to sky to completely
freeze. Defaults to False

CooledLavaBlock: Block used as cooled or frozen lava. Set this to
OBSIDIAN for "frozen" lava lakes in cold biomes. Defaults to
STATIONARY_LAVA. NOTE: This has a setting in both World and Biome configs
and follows the same behavior as WaterBlock and IceBlock when using
the UseWorldWaterLevel setting.