Remove old biome registration hack
The CraftBlock BiomeBase <--> Biome mappings are no longer there, so
there is no reason to even try updating them.

This does mean that plugins will get an IllegalArgumentException when
getting a custiom biome using the Bukkit API. To fix this, we would need
to add an enum constant to the Biome enum. Forge has an EnumHelper class
for this, we probably need to include this with TC.

Also fixed the Logger sometimes printing empty lines.