Fix Eye of Enders not pointing towards strongholds
To fix this, StrongholdGen had to extend WorldGenStronghold, otherwise
it couldn't be injected.

ReflectionHelper has been made more strict: it will now only pick up
fields if the field is the only one of its type in the class. This will
avoid silent errors when a new field of the same type is added to a
class. (Like a second List to WorldGenStronghold.)