Don't automatically sort resources queue
This is a partial revert of e0abc791452da1d06c88b089e55936694bcef8e8 .
The reason for this is that it is now possible again to, for example,
spawn a specific BO3, and then place some trees over it.

Not everything is reverted. The resources in the default settings are
still sorted according to their natural order. In old versions of TC,
this order depended on the order of the code in createDefaultResources.
This behaviour is not restored.