Fix another bug in ReplaceToBiomeName
When ReplaceToBiomeName was set to a biome loaded after this biome, or
when ReplaceToBiomeName was set to a biome that also has a
ReplaceToBiomeName value, the server would crash.

To correctly cross-reference settings between biomes, all biomes need to
be loaded first, and only then they may look at the settings of each
other.