TC console commands (#407)
Here's a class that adds console commands for Forge. Just add this to TCPLugin to hook it up:

    @EventHandler
    public void serverLoad(FMLServerStartingEvent event)
    {
    	event.registerServerCommand(new TCCommandHandler());
    }

You may want to add "Author" and "Description" settings to the world config as well for the /worldinfo command. If you could finish the /tc biome command that'd be nice! Rename this file or move it wherever you want ofc.