Add world creation menu
This replaces the world creation menu with a custom world creation menu. Actually it also adds a menu and makes changes to the vanilla world selection menu so I guess this file's name isn't entirely accurate but what the heck. Rename it and put it anywhere you like. To hook this up add this to TCPlugin:

MinecraftForge.EVENT_BUS.register(new WorldCreationMenuHandler());