Make all ranges inclusive; rename canyons to ravines
Canyons are now called ravines. Settings are converted automatically.
This should make the settings easier to find for new users, as Minecraft
calls it ravines too.

# Ranges
As far as I know, as of this commit, all ranges (like
MinAltitude/MaxAltitude, or MinHeight and MaxHeight in BO3s) in
TerrainControl are now inclusive. This means that it is now always
possible to set the max value to the same value as the min value.

This fixes an annoying bug. Previously, there were situations where the
terrain generator treated a range as inclusive but the settings checker
treated the range as exclusive. MinHeight and MaxHeight in BO3s were one
example of this: you couldn't set MaxHeight to the same value as
MinHeight, so you couldn't pin a BO3 to a single height. SpawnHeight:
atMinY was added as a workaround for this.

## Effect on WorldConfig
When TC detects that you still have canyon settings in your config, it
knows that you still had the old exclusive ranges for the following
settings:

* CaveMaxAltitude
* CaveSystemPocketMaxSize
* RavineMaxAltitude
* RavineMaxLength

It will then automatically lower the values of those settings by one.

## Effect on biome configs
The following resources were still using exclusive ranges, and will now
use inclusive ranges:

* Cactus
* Dungeon
* Liquid
* Ore
* Plant
* SmallLake
* UndergroundLake
* Vines

Unlike for the WorldConfig, the values in those resources are not
automatically updated. This will cause some small differences in terrain
generation.