Introduce SettingsMap
Previously, settings were read by a SettingsReader and written to a
SettingsWriter.

The SettingsReader and SettingsWriter interfaces were very similar: the
SettingsReader interface allowed you to write settings to memory, to
handle outdated config formats.

Both interfaces have been combined into one interface, called
SettingsMap.

In SettingsMap, each comment must belong to a setting or a header. This
design has several benefits. First, it might be possbile to read
comments in the future, so that custom comments are kept. Second, for
biome inheritance, it is now possible to write only parts of a config
file, as we now know what a comment is attached to.

None of those things is implemented yet, but probably will be in the
future.

The BiomeConfig class has also been a bit simplified; it now reads the
settings in the constructor. All the different states we had before are
gone now.

BO3s, BiomeConfigs and the WorldConfig no longer keep their raw settings
in memory. This shouuld save quite some RAM for worlds with lots of BO3s
or biomes.