Make ConfigFunction code easier to use
## Instance state and the init method
Previously, there was all kinds of complex state in the ConfigFunction
instance. Now all required arguments are provided in the constructor of
a ConfigFunction. The init method is gone.

## Error handling
Previously, if the init method threw an InvalidConfigException, the
instance would be kept. Now such invalid instances are no longer kept,
an instance of ErroredFunction is used instead. The getError and isValid
methods have been removed.

## Reflection code
There's still some relfection code involved, as we need a way to go from
for example PlantGen.class to a PlantGen instance. If we would use Java
8, we could use the newer PlantGen::new syntax to get a reference to the
constructor of PlantGen.

## makeString()
The ConfigFunction.makeString() method has been renamed to the more
conventional toString(). The ConfigFunction.write() wrapper method has
been removed.

## getHolderType()
This method has been removed. Checking for the holder type now happens
by checking for which constructors exist.