Remove BiomeConfig.generationId
For Forge, we'll need to allow Forge to assign an id to us. A final
generationId field in BiomeConfig won't be reliable anymore.

Forge biome registration has changed a bit to be more in line with how
Forge wants us to register things, but a singleplayer world still
crashes when loaded for a second time.